Extract Json related functions to a separate utility class
